--- app/src/main/java/com/example/evsherpa/ui/subsidyStatus/SubsidyStatusFragmentImg.java
@@ -1,22 +1,0 @@
-package com.example.evsherpa.ui.subsidyStatus;
-
-import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-import androidx.annotation.NonNull;
-
-import com.example.evsherpa.R;
-
-public class SubsidyStatusFragmentImg {
-
-    public View onCreateView(@NonNull LayoutInflater inflater,
-                             ViewGroup container, Bundle savedInstanceState) {
-
-        View view = inflater.inflate(R.layout.fragment_subsidy_status, container, false);
-
-
-        return view;
-    }
-}
